Added a new option on ResultLink
titleTemplate: allows you to define a display template for what will be displayed as text for the ResultLink. It supports the same syntax as the hrefTemplate and behaves the same way.

Modified the method parseHrefTemplate and made it into the more generic parseStringTemplate. The same method is now used for both hrefTemplate and titleTemplate.

Added a logger warning when using this template parser for when the key entered is undefined.

Added the corresponding tests.