Linked 2 components
All the rendering for the resultLink is now done by the renderUri function, just like the PrintableUri. ResultLink initiates the render instead of PrintableUri. Since PrintableUri already has a renderUri function, it is the one that is used for rendering. Need to add code to handle added options from ResultLink.
https://coveord.atlassian.net/browse/JSUI-1536