updated unit tests
Used more precise selectors as the text from the new "title" element was getting appended to the textContent/innerText propperty of the HTMLElement.

In some tests, the text wanted was part of the current selector, but the text from its child (svg tag) was getting append as well. So, I removed the 'svg' children of the parent in order to have only the original text in the textContent/innerHTML tag

https://coveord.atlassian.net/browse/JSUI-3058